[PATCH] s390: core s390 update
From: Martin Schwidefsky <schwidefsky@de.ibm.com>

s390 bug fixes:

- Add missing i/o controls to compat ioctl translation table.

- Fix some gcc 3.4 warnings.

- Export _sb_findmap.

- Export smp_call_function_on only if CONFIG_SMP=y.

- Add safe-guard to diag10.

- Add type definition for compat_timer_t.

- Fix first argument of signal_processor_ps.

+++ include/asm-s390/sigp.h
@@ -121,7 +121,7 @@ signal_processor_p(unsigned long parameter,__u16 cpu_addr,
  * Signal processor with parameter and return status
  */
 extern __inline__ sigp_ccode
-signal_processor_ps(__u32 *statusptr, unsigned long parameter,
+signal_processor_ps(unsigned long *statusptr, unsigned long parameter,
 		    __u16 cpu_addr, sigp_order_code order_code)
 {
 	sigp_ccode ccode;
